middy-custom-middleware: Change on folder structure
